utils/pdf_report.py
# Sinamor
from fpdf import FPDF

def generate_pdf_report(session_summary):
    pdf = FPDF()
    pdf.add_page()
    pdf.set_font("Arial", size=12)
    pdf.cell(200, 10, txt="MindMitra - Mental Health Report", ln=True, align='C')
    # More report logic to come
    return pdf.output(dest='S').encode('latin-1')
