requirements.txt
# Sinamor
streamlit
transformers
torch
plotly
pandas
SpeechRecognition
pyttsx3
fpdf2
